+++ src/calendar.binding.js
@@ -4,7 +4,7 @@
  * A javascript drag & drop event calendar, that is fully responsive and compatible with all modern browsers.
  * 
  * @file        calendar.binding.js
- * @version     v2.3.1
+ * @version     v2.3.2
  * @author      Bunoon
  * @license     GNU AGPLv3
  * @copyright   Bunoon 2023