+++ src/calendar.js
@@ -4,7 +4,7 @@
  * A javascript drag & drop event calendar, that is fully responsive and compatible with all modern browsers.
  * 
  * @file        calendar.js
- * @version     v2.3.1
+ * @version     v2.3.2
  * @author      Bunoon
  * @license     GNU AGPLv3
  * @copyright   Bunoon 2023
@@ -2140,10 +2140,15 @@ function calendarJs( elementOrId, options, searchOptions ) {
     }
 
     function hideYearSelectorDropDown() {
+        var closed = false;
+
         if ( isYearSelectorDropDownVisible() ) {
             _element_HeaderDateDisplay_YearSelector_DropDown_Arrow.className = "ib-arrow-down-full-medium";
             _element_HeaderDateDisplay_YearSelector.style.display = "none";
+            closed = true;
         }
+
+        return closed;
     }
 
     function isYearSelectorDropDownVisible() {
@@ -2179,11 +2184,11 @@ function calendarJs( elementOrId, options, searchOptions ) {
             windowFunc = addEvents ? _window.addEventListener : _window.removeEventListener;
 
         documentBodyFunc( "click", onDocumentClick );
-        documentBodyFunc( "contextmenu", hideAllDropDowns );
+        documentBodyFunc( "contextmenu", onEventHideAllDropDowns );
         documentBodyFunc( "mousemove", onMoveDocumentMouseMove );
         documentBodyFunc( "mouseleave", onMoveDocumentMouseLeave );
-        documentFunc( "scroll", hideAllDropDowns );
-        windowFunc( "resize", hideAllDropDowns );
+        documentFunc( "scroll", onEventHideAllDropDowns );
+        windowFunc( "resize", onEventHideAllDropDowns );
         windowFunc( "resize", centerSearchDialog );
         windowFunc( "resize", onWindowResizeRefreshViews );
         windowFunc( "blur", onWindowFocusOut );
@@ -2218,14 +2223,42 @@ function calendarJs( elementOrId, options, searchOptions ) {
         }, 50, false );
     }
 
-    function hideAllDropDowns() {
-        hideDropDownMenu( _element_DropDownMenu_Day );
-        hideDropDownMenu( _element_DropDownMenu_Event );
-        hideDropDownMenu( _element_DropDownMenu_FullDay );
-        hideDropDownMenu( _element_DropDownMenu_HeaderDay );
-        hideSearchHistoryDropDownMenu();
-        hideYearSelectorDropDown();
+    function onEventHideAllDropDowns() {
+        hideAllDropDowns();
+    }
+
+    function hideAllDropDowns( hideSearchHistoryDropDown ) {
+        var itemClosed = false;
+
+        hideSearchHistoryDropDown = isDefined( hideSearchHistoryDropDown ) ? hideSearchHistoryDropDown : true;
+
+        if ( hideDropDownMenu( _element_DropDownMenu_Day ) ) {
+            itemClosed = true;
+        }
+
+        if ( hideDropDownMenu( _element_DropDownMenu_Event ) ) {
+            itemClosed = true;
+        }
+
+        if ( hideDropDownMenu( _element_DropDownMenu_FullDay ) ) {
+            itemClosed = true;
+        }
+
+        if ( hideDropDownMenu( _element_DropDownMenu_HeaderDay ) ) {
+            itemClosed = true;
+        }
+
+        if ( hideYearSelectorDropDown() ) {
+            itemClosed = true;
+        }
+
         hideTooltip();
+
+        if ( hideSearchHistoryDropDown ) {
+            hideSearchHistoryDropDownMenu();
+        }
+
+        return itemClosed;
     }
 
     function onWindowKeyDown( e ) {
@@ -2361,12 +2394,13 @@ function calendarJs( elementOrId, options, searchOptions ) {
     }
 
     function closeActiveDialog() {
-        var done = false;
+        var done = hideAllDropDowns( false );
 
-        hideAllDropDowns();
-        clearSelectedEvents();
+        if ( !done ) {
+            done = hideSearchHistoryDropDownMenu();
+        }
 
-        if ( _openDialogs.length > 0 ) {
+        if ( !done && _openDialogs.length > 0 ) {
             var lastFunc = _openDialogs[ _openDialogs.length - 1 ];
 
             if ( isFunction( lastFunc ) ) {    
@@ -2379,6 +2413,10 @@ function calendarJs( elementOrId, options, searchOptions ) {
 
         if ( !done ) {
             done = hideSearchDialog();
+        }
+
+        if ( !done ) {
+            done = clearSelectedEvents();
         }
 
         if ( !done && _copiedEventDetails.length > 0 ) {
@@ -5627,9 +5665,14 @@ function calendarJs( elementOrId, options, searchOptions ) {
     }
 
     function hideDropDownMenu( element ) {
+        var closed = false;
+
         if ( isDropDownMenuVisible( element ) ) {
             element.style.display = "none";
+            closed = true;
         }
+
+        return closed;
     }
 
     function isDropDownMenuVisible( element ) {
@@ -7262,14 +7305,20 @@ function calendarJs( elementOrId, options, searchOptions ) {
     }
 
     function hideSearchHistoryDropDownMenu() {
-        if ( _element_SearchDialog_History_DropDown !== null ) {
+        var closed = false;
+
+        if ( _element_SearchDialog_History_DropDown !== null && _element_SearchDialog_History_DropDown_Button.className == "ib-arrow-up-full" ) {
             _element_SearchDialog_History_DropDown.style.display = "none";
             _element_SearchDialog_History_DropDown_Button.className = "ib-arrow-down-full";
+
+            closed = true;
         }
+
+        return closed;
     }
 
     function showSearchHistoryDropDownMenu() {
-        if ( _element_SearchDialog_History_DropDown !== null ) {
+        if ( _element_SearchDialog_History_DropDown !== null && _element_SearchDialog_History_DropDown_Button.className == "ib-arrow-down-full" ) {
             _element_SearchDialog_History_DropDown.style.display = "block";
             _element_SearchDialog_History_DropDown_Button.className = "ib-arrow-up-full";
         }
@@ -8060,13 +8109,19 @@ function calendarJs( elementOrId, options, searchOptions ) {
     }
 
     function clearSelectedEvents() {
-        var eventsSelectedLength = _eventsSelected.length;
+        var cleared = false,
+            eventsSelectedLength = _eventsSelected.length;
 
-        for ( var eventsSelectedIndex = 0; eventsSelectedIndex < eventsSelectedLength; eventsSelectedIndex++ ) {
-            updateEventClasses( _eventsSelected[ eventsSelectedIndex ].id, "selected-event", true );
+        if ( eventsSelectedLength > 0 ) {
+            for ( var eventsSelectedIndex = 0; eventsSelectedIndex < eventsSelectedLength; eventsSelectedIndex++ ) {
+                updateEventClasses( _eventsSelected[ eventsSelectedIndex ].id, "selected-event", true );
+            }
+    
+            cleared = true;
+            _eventsSelected = [];
         }
 
-        _eventsSelected = [];
+        return cleared;
     }
 
     function setCopiedEventsFromKeyDown( cut ) {
@@ -9815,6 +9870,14 @@ function calendarJs( elementOrId, options, searchOptions ) {
         clearElementsByClassName( _document.body, "calendar-tooltip-event" );
         clearElementsByClassName( _document.body, "calendar-notification" );
 
+        if ( _datePickerModeEnabled ) {
+            _document.removeEventListener( "click", hideDatePickerMode );
+        }
+
+        if ( _options.tooltipsEnabled ) {
+            document.body.removeEventListener( "mousemove", hideTooltip );
+        }
+
         _element_Calendar.className = _string.empty;
         _element_Calendar.innerHTML = _string.empty;
 
@@ -10907,7 +10970,7 @@ function calendarJs( elementOrId, options, searchOptions ) {
      * @returns     {string}                                                The version number.
      */
     this.getVersion = function() {
-        return "2.3.1";
+        return "2.3.2";
     };
 
     /**